+++ third-party/svg-sanitize/src/Exceptions/NestingException.php
@@ -1,0 +1,40 @@
+<?php
+namespace WPVibe\Vendor\enshrined\svgSanitize\Exceptions;
+
+if ( ! defined( 'ABSPATH' ) ) {
+    exit;
+}
+
+use Exception;
+
+class NestingException extends \Exception
+{
+    /**
+     * @var \DOMElement
+     */
+    protected $element;
+
+    /**
+     * NestingException constructor.
+     *
+     * @param string           $message
+     * @param int              $code
+     * @param Exception|null   $previous
+     * @param \DOMElement|null $element
+     */
+    public function __construct($message = "", $code = 0, ?Exception $previous = null, ?\DOMElement $element = null)
+    {
+        $this->element = $element;
+        parent::__construct($message, $code, $previous);
+    }
+
+    /**
+     * Get the element that caused the exception.
+     *
+     * @return \DOMElement
+     */
+    public function getElement()
+    {
+        return $this->element;
+    }
+}